# server/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "description": "This is a swagger documentation for decision engine api",
        "version": "0.0.1",
        "title": "Decision Engine API",
        "termsOfService": "https://smartbear.com/terms-of-use/",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
...
    "paths": {
        "/api/users/signIn": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign In",
                "description": "Sign In",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            },
                            "example": {
                                "email": "test@gmail.com",
                                "password": "Test123"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid request parameters",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResp"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResp"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
...
        "/api/datamodel/all/{companyId}": {
            "get": {
                "tags": [
                    "DataModel"
                ],
                "summary": "Get all datamodels by company",
                "description": "Get all datamodels by company",
                "parameters":[
                    {
                    "in":"path",
                    "name":"companyId",
                    "required":true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "*/*" : {
                                "schema": {
                                    "$ref": "#/components/schemas/ListDataModelResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid request parameters",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResp"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResp"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SignInRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "SignInResponse": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "token_version": {
                        "type": "number"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdDate": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            },
...
            "ListDataModelResponse":{
                "type": "object",
                "properties": {
                    "results": {
                        "type": "array",
                        "properties": {
                            "item": {
                                "type":"object",
                                "properties":{
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "prefix": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "createdBy": {
                                        "type": "object"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "properties": {
                                        "type": "object"
                                    },
                                    "isActive": {
                                        "type": "boolean"
                                    },
                                    "createdAt": {
                                        "type": "string"
                                    },
                                    "updatedAt": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "hasNextPage": {
                        "type": "boolean"
                    },
                    "totalPages": {
                        "type": "number"
                    },
                    "totalResults": {
                        "type": "number"
                    }
                }
            },
...
            "ErrorResp": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "field": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}